Started on ACS 1996 drawing mode. Significant change (not by itself, sadly) is that MolDrawOptions::lineWidth has changed from int to double to allow for ACS requirement of 0.6px bond widths.

+++ Code/GraphMol/MolDraw2D/DrawMol.h
@@ -170,8 +170,8 @@ class DrawMol {
   std::pair<DrawColour, DrawColour> getBondColours(Bond *bond);
   void makeContinuousHighlights();
   void makeAtomCircleHighlights();
-  void makeAtomEllipseHighlights(int lineWidth);
-  void makeBondHighlightLines(int lineWidth);
+  void makeAtomEllipseHighlights(double lineWidth);
+  void makeBondHighlightLines(double lineWidth);
   void calcAnnotationPosition(const Atom *atom,
                               DrawAnnotation &annot) const;
   void calcAnnotationPosition(const Bond *bond,
